third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-request-xhr.html

<!DOCTYPE html>
<title>Service Worker: the body of FetchEvent using XMLHttpRequest</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js?pipe=sub"></script>
<script>
async_test(function(t) {
    var SCOPE = 'resources/fetch-request-xhr-iframe.html';
    var SCRIPT = 'resources/fetch-request-xhr-worker.js';
    var host_info = get_host_info();
    service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(function(sw) {
          return wait_for_state(t, sw, 'activated');
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(frame) {
          var channel = new MessageChannel();
          channel.port1.onmessage = t.step_func(function(e) {
              assert_equals(e.data.results, 'finish');
              unload_iframe(frame);
              service_worker_unregister_and_done(t, SCOPE);
            });
          frame.contentWindow.postMessage({},
                                          [channel.port2],
                                          host_info['HTTP_ORIGIN']);
        })
      .catch(unreached_rejection(t));
  }, 'Verify the body of FetchEvent using XMLHttpRequest');
</script>
